com.github.AppDevNext.Logcat:LogcatCoreLib library deleted due to compatibility issues with the app tests.
