+++ gradle/libs.versions.toml
@@ -24,7 +24,6 @@ androidxWork = "2.8.1"
 coil = "2.2.2"
 dexopener = "2.0.5"
 disklrucache = "2.0.2"
-logcatcorelib = "3.2"
 media3 ="1.1.1"
 floatingactionbutton = "1.10.1"
 glide = "4.15.1"
@@ -44,6 +43,7 @@ photoView = "2.3.0"
 preference = "1.2.0"
 sonarqube = "4.0.0.2929"
 stetho = "1.6.0"
+timber = "5.0.1"
 
 [libraries]
 androidx-annotation = { group = "androidx.annotation", name = "annotation", version.ref = "androidxAnnotation" }
@@ -79,7 +79,6 @@ androidx-test-rules = { group = "androidx.test", name = "rules", version.ref = "
 androidx-test-runner = { group = "androidx.test", name = "runner", version.ref = "androidxTest" }
 androidx-test-uiautomator = { group = "androidx.test.uiautomator", name = "uiautomator", version.ref = "androidxTestUiAutomator" }
 androidx-work-runtime-ktx = { group = "androidx.work", name = "work-runtime-ktx", version.ref = "androidxWork" }
-appdevnext-logcat-logcatcorelib = { group = "com.github.AppDevNext.Logcat", name = "LogcatCoreLib", version.ref = "logcatcorelib" }
 coil = { group = "io.coil-kt", name = "coil", version.ref = "coil" }
 dexopener = { group = "com.github.tmurakami", name = "dexopener", version.ref = "dexopener" }
 disklrucache = { group = "com.jakewharton", name = "disklrucache", version.ref = "disklrucache" }
@@ -109,6 +108,7 @@ moshi-kotlin-codegen = { group = "com.squareup.moshi", name = "moshi-kotlin-code
 patternlockview = { group = "com.github.aritraroy.PatternLockView", name = "patternlockview", version.ref = "patternlockview" }
 photoview = { group = "com.github.chrisbanes", name = "PhotoView", version.ref = "photoView" }
 stetho = { group = "com.facebook.stetho", name = "stetho", version.ref = "stetho" }
+timber = { group = "com.jakewharton.timber", name = "timber", version.ref = "timber" }
 
 # Dependencies of the included build-logic
 android-gradlePlugin = { group = "com.android.tools.build", name = "gradle", version.ref = "androidGradlePlugin" }
